Revert "feat: add icons to autocomplete"
I noticed when saving that an autocommand was running that was applying
a gray bar seemingly random in the editor which was very annoying. This
bar would overwrite characters. I wasn't able to determine if this bug
was the result of adding this new package or update a whole bunch of
packages at once. Next time I will need to be more careful with updating
dependencies.

This reverts commit bb89cf648c4440ea506e830c1dfb529c26bf6426.